New submodules for AXI and updated vcompile scripts and IP wrappers for new axi_node and so on

+++ ips/scripts/vcompile_common_cells.sh
@@ -3,6 +3,18 @@ source scripts/colors.sh
 
 echo "${Green}--> Compiling pulp_common_cells... ${NC}"
 
+# check if environment variables are defined
+if (! $?MSIM_LIBS_PATH ) then
+  echo "${Red} MSIM_LIBS_PATH is not defined ${NC}"
+  exit 1
+endif
+
+if (! $?IPS_PATH ) then
+  echo "${Red} IPS_PATH is not defined ${NC}"
+  exit 1
+endif
+
+
 echo "${Green}library: pulp_common_cells ${NC}"
 rm -rf ${MSIM_LIBS_PATH}/pulp_common_cells_lib
 
@@ -12,52 +24,22 @@ vmap pulp_common_cells_lib ${MSIM_LIBS_PATH}/pulp_common_cells_lib
 echo "${Green}Compiling component:   ${Brown} pulp_common_cells ${NC}"
 echo "${Red}"
 
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_buffer.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_inverter.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_gating.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_mux2.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_xor2.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_buffer.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_in.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_in_clamp.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_out.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_out_clamp.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_and2.sv
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_buffer.sv      || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_inverter.sv    || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_gating.sv      || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_mux2.sv        || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_xor2.sv        || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_buffer.sv            || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_and2.sv        || exit 1
 
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_clock_buffer_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_clock_inverter_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_clock_gating_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_clock_mux2_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_clock_xor2_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_in_cmos028fdsoi.sv
-# //vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/pulp_level_shifter_out_cmos028fdsoi.sv
 
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_buffer.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_inverter.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_gating.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_mux2.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_xor2.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_in.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_in_clamp.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_out.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_out_clamp.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_and2.sv
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_buffer.sv   || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_inverter.sv || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_gating.sv   || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_mux2.sv     || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_xor2.sv     || exit 1
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/cluster_clock_and2.sv     || exit 1
 
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_clock_buffer_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_clock_inverter_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_clock_gating_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_clock_mux2_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_clock_xor2_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_in_cmos028fdsoi.sv
-#//vlog -sv -work pulp_common_cells_lib +incdir+../rtl/includes -quiet ${IPS_PATH}/common_cells/cluster_level_shifter_out_cmos028fdsoi.sv
-
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/edge_propagator_tx.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/edge_propagator_rx.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/edge_propagator.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_sync_wedge.sv
-
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_clock_gating_async.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/pulp_sync_wedge.sv
-vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/rstgen.sv
+vlog -sv -work pulp_common_cells_lib -quiet ${IPS_PATH}/common_cells/generic_fifo.sv           || exit 1
 
 echo "${Cyan}--> pulp_common_cells compilation complete! ${NC}"
